Try the other solution aturon and I came up with:
    0: STORE_sc(x)
    | sequenced-before, sc
    1: READ_sc(y)
    | sc
    2: CAS_sc(y)
    | sequenced-before, sc
    3: LOAD_sc(x)

Here we use sc everywhere. The reason for this is that we wish to
guarantee that 3 observes 0. For this, acq/rel are not sufficient: they
would only say "if 3 observes 0, then it observes whatever happens
before 0 too", but we wish to guarantee that if 1-sc->2, then 0-hb->3.
Using sc ensures that.